Refactor: Rename modbussetter and viewcallback files.
